Rename StoredResources.h to StorableResources.h
