+++ OPHD/StructureCatalogue.h
@@ -5,7 +5,7 @@
 
 #include "Things/Structures/Structures.h"
 
-#include "StoredResources.h"
+#include "StorableResources.h"
 
 
 /**
