+++ OPHD/ophd.vcxproj
@@ -333,7 +333,7 @@ IF NOT "$(VcpkgCurrentInstalledDir)" == "" (
     <ClInclude Include="Map\TileMap.h" />
     <ClInclude Include="MicroPather\micropather.h" />
     <ClInclude Include="Mine.h" />
-    <ClInclude Include="StoredResources.h" />
+    <ClInclude Include="StorableResources.h" />
     <ClInclude Include="PopulationPool.h" />
     <ClInclude Include="Population\Morale.h" />
     <ClInclude Include="Population\Population.h" />
