+++ OPHD/ophd.vcxproj.filters
@@ -635,7 +635,7 @@
     <ClInclude Include="States\Route.h">
       <Filter>Header Files\States</Filter>
     </ClInclude>
-    <ClInclude Include="StoredResources.h">
+    <ClInclude Include="StorableResources.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
